+++ app/Http/Controllers/API/PurchaseController.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Http\Controllers\API;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+class PurchaseController
+{
+    //
+}
