+++ app/Http/Requests/Attribute/AttributeRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Attribute;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StockAdjustmentRequest extends FormRequest
+class AttributeRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,7 +22,7 @@ class StockAdjustmentRequest extends FormRequest
     public function rules(): array
     {
         return [
-            //
+            'name' => 'required|max:50'
         ];
     }
 }
