+++ app/Http/Requests/Product/ProductAttributeRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Product;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class AttributesValueRequest extends FormRequest
+class ProductAttributeRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -21,11 +21,8 @@ class AttributesValueRequest extends FormRequest
      */
     public function rules(): array
     {
-        
-        $id = $this->route('id');
         return [
-          //  'attribute_id' => 'required',
-            'value' => 'required|unique:attributes_values,value'
+            'attribute_value_id' => 'required'
         ];
     }
 }
