+++ app/Http/Requests/Stocks/StockMovementRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Stocks;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
